requirements.txt
flask
pyjwt
Flask-SQLAlchemy
flask-restful
gunicorn
lyricsgenius
configparser